// semant_demo_frontend/src/generated/api/models/SpanType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SpanType = {
    pos: 'pos',
    neg: 'neg',
    auto: 'auto'
} as const;
export type SpanType = typeof SpanType[keyof typeof SpanType];


export function instanceOfSpanType(value: any): boolean {
    for (const key in SpanType) {
        if (Object.prototype.hasOwnProperty.call(SpanType, key)) {
            if (SpanType[key as keyof typeof SpanType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SpanTypeFromJSON(json: any): SpanType {
    return SpanTypeFromJSONTyped(json, false);
}

export function SpanTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): SpanType {
    return json as SpanType;
}

export function SpanTypeToJSON(value?: SpanType | null): any {
    return value as any;
}

export function SpanTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): SpanType {
    return value as SpanType;
}