// semant_demo_frontend/src/generated/api/models/WriteOutcome.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const WriteOutcome = {
    complete: 'complete',
    partial: 'partial',
    failed: 'failed'
} as const;
export type WriteOutcome = typeof WriteOutcome[keyof typeof WriteOutcome];


export function instanceOfWriteOutcome(value: any): boolean {
    for (const key in WriteOutcome) {
        if (Object.prototype.hasOwnProperty.call(WriteOutcome, key)) {
            if (WriteOutcome[key as keyof typeof WriteOutcome] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WriteOutcomeFromJSON(json: any): WriteOutcome {
    return WriteOutcomeFromJSONTyped(json, false);
}

export function WriteOutcomeFromJSONTyped(json: any, ignoreDiscriminator: boolean): WriteOutcome {
    return json as WriteOutcome;
}

export function WriteOutcomeToJSON(value?: WriteOutcome | null): any {
    return value as any;
}

export function WriteOutcomeToJSONTyped(value: any, ignoreDiscriminator: boolean): WriteOutcome {
    return value as WriteOutcome;
}